Add http/1.x header module
